Refactor lines models
