+++ server/apps/lines/models/__init__.py
@@ -1,0 +1,2 @@
+from .lines import LineModel
+from .routes import RouteModel
