+++ server/apps/lines/models/routes.py
@@ -1,28 +1,20 @@
 # coding: utf-8
 from django.db import models
 
+from apps.lines.models import LineModel
 from apps.stations.models import StationModel
 
 from apps.utils import create_id
-
-
-def create_line_id():
-    return create_id('line_')
 
 
 def create_route_id():
     return create_id('route_')
 
 
-class LineModel(models.Model):
-
-    id = models.CharField(default=create_line_id, primary_key=True,
-                          max_length=30, unique=True)
-    name = models.CharField(max_length=100)
-    color = models.CharField(max_length=8)
-
-
 class RouteModel(models.Model):
+
+    class Meta:
+        ordering = ['-id']
 
     id = models.CharField(default=create_route_id, primary_key=True,
                           max_length=30, unique=True)
